correct-crud

+++ controller/generic/generic.js
@@ -7,7 +7,16 @@ const findAll = (req, res) => {
 
     const reqBody = req.body
 
-    crud.findAll(req.entity, db, reqBody).then(genres => {
+    // get criteria from query parameters
+    const base64Criteria = req.query.criteria
+    let criteria = {}
+    try {
+        criteria = base64Criteria ? JSON.parse(Buffer.from(base64Criteria, 'base64').toString('ascii')) : {}
+    } catch (e) {
+        //     do nothing
+    }
+
+    crud.findAll(req.entity, db, criteria).then(genres => {
             res.send(genres)
         }
     )
@@ -37,30 +46,48 @@ const create = (req, res) => {
 const update = (req, res) => {
     checkEntity(req, res)
     let id = req.params[0]
-    crud.update(req.entity, req.db,req.body, id).then(result => {
+    crud.update(req.entity, req.db, req.body, id).then(result => {
             // newly created object
-            result==null || result.value === null ? sendError(res, 'No document found', 500) :
-            res.send(result.value)
+            result == null || result.value === null ? sendError(res, 'No document found', 500) :
+                res.send(result.value)
         }
     )
         .catch(error => sendError(res, error, 500))
 }
 
-const deleteOne = (req, res) => {
+const deleteOne = async (req, res) => {
     checkEntity(req, res)
     let id = req.params[0]
+
     // Find the document before deleting
-    crud.deleteOne(req.entity, req.db, id).then(result => {
-        if (result.deletedCount === 0) {
-            sendError(res, 'No document to delete', 500)
-        }
-        res.send({
-                code: 200,
-                message: "Document deleted",
-                id: id
+    const element = await crud.findOne(req.entity, req.db, id)
+    if (element === null) {
+        sendError(res, 'No document to delete', 500)
+    } else if (element.status === undefined) {
+        // Find the document before deleting
+        crud.deleteOne(req.entity, req.db, id).then(result => {
+            if (result.deletedCount === 0) {
+                sendError(res, 'No document to delete', 500)
             }
-        )         
-    }) .catch(error => sendError(res, error, 500));
+            res.send({
+                    code: 200,
+                    message: "Document deleted",
+                    id: id
+                }
+            )
+        }).catch(error => sendError(res, error, 500));
+    } else if (element.status !== undefined) {
+        element.status = 0
+        crud.update(req.entity, req.db, element, id).then(result => {
+            // newly created object
+            result == null || result.value === null ? sendError(res, 'No document found', 500) :
+                res.send({
+                    code: 200,
+                    message: "Document deleted",
+                    id: id
+                })
+        }).catch(error => sendError(res, error, 500));
+    }
 }
 
 const checkEntity = (req, res) => {
